feat: update coding standards and related modules
- Updated CODING_STYLE.md with wheel loader specific guidelines
- Enhanced WheelSpeedsSetpoint.msg with proper documentation
- Improved AS5600 magnetic encoder driver
- Updated slip estimator module
- Enhanced steering controller with proper coding standards
- Improved terrain adaptation module
- Updated wheel loader controller main module

All changes follow PX4 coding standards and wheel loader project conventions.

+++ src/modules/articulated_chassis/slip_estimator/slip_estimator.cpp
@@ -66,7 +66,7 @@ void SlipEstimator::updateVehicleState()
             // Verify this is the correct instance
             if (encoder_data.instance == front_idx) {
                 // Convert 1/million rad/s to linear speed (m/s) using wheel radius parameter
-                _vehicle_state.wheel_speed_front = encoder_data.velocity * 1e-6f * _param_wheel_radius.get();
+                _vehicle_state.wheel_speed_front = static_cast<float>(encoder_data.velocity * 1e-6) * _param_wheel_radius.get();
             }
         }
     }
@@ -78,7 +78,7 @@ void SlipEstimator::updateVehicleState()
             // Verify this is the correct instance
             if (encoder_data.instance == rear_idx) {
                 // Convert 1/million rad/s to linear speed (m/s) using wheel radius parameter
-                _vehicle_state.wheel_speed_rear = encoder_data.velocity * 1e-6f * _param_wheel_radius.get();
+                _vehicle_state.wheel_speed_rear = static_cast<float>(encoder_data.velocity * 1e-6) * _param_wheel_radius.get();
                 _vehicle_state.last_update = encoder_data.timestamp;
             }
         }
